Added resume action, when submission is Rejected and an upload occurs
